+++ CardForm/src/main/java/com/braintreepayments/api/SupportedCardTypesAdapter.java
@@ -1,0 +1,77 @@
+package com.braintreepayments.api;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.VisibleForTesting;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.braintreepayments.api.cardform.R;
+
+class SupportedCardTypesAdapter extends RecyclerView.Adapter<SupportedCardTypesAdapter.SupportedCardTypesViewHolder> {
+
+    private final int OPAQUE = 255;
+    private final int SEMI_TRANSPARENT = 80;
+
+    private final SelectableCardType[] supportedCardTypes;
+
+    static class SupportedCardTypesViewHolder extends RecyclerView.ViewHolder {
+        private final ImageView imageView;
+
+        public SupportedCardTypesViewHolder(View view) {
+            super(view);
+            imageView = view.findViewById(R.id.bt_supported_card_icon);
+        }
+
+        public ImageView getImageView() {
+            return imageView;
+        }
+    }
+
+    SupportedCardTypesAdapter(SelectableCardType[] cardTypes) {
+        supportedCardTypes = cardTypes;
+    }
+
+    @NonNull
+    @Override
+    public SupportedCardTypesViewHolder onCreateViewHolder(ViewGroup viewGroup, int viewType) {
+        View view = LayoutInflater.from(viewGroup.getContext())
+                .inflate(R.layout.bt_supported_card_type, viewGroup, false);
+
+        return new SupportedCardTypesViewHolder(view);
+    }
+
+    @Override
+    public void onBindViewHolder(SupportedCardTypesViewHolder viewHolder, final int position) {
+        SelectableCardType selectableCardType = supportedCardTypes[position];
+        viewHolder.getImageView().setImageResource(selectableCardType.getCardType().getFrontResource());
+
+        viewHolder.getImageView().setContentDescription(selectableCardType.getCardType().toString());
+        if (selectableCardType.isDisabled()) {
+            viewHolder.getImageView().setImageAlpha(SEMI_TRANSPARENT);
+        } else {
+            viewHolder.getImageView().setImageAlpha(OPAQUE);
+        }
+    }
+
+    @Override
+    public int getItemCount() {
+        return supportedCardTypes.length;
+    }
+
+    @VisibleForTesting
+    SelectableCardType[] getSupportedCardTypes() {
+        return supportedCardTypes;
+    }
+
+    void setSelected(CardType cardType) {
+        if (supportedCardTypes != null) {
+            for (SelectableCardType selectableCardType : supportedCardTypes) {
+                selectableCardType.setDisabled(selectableCardType.getCardType() != cardType);
+            }
+        }
+    }
+}