+++ CardForm/src/test/java/com/braintreepayments/api/PostalCodeEditTextTest.java
@@ -1,0 +1,50 @@
+package com.braintreepayments.api;
+
+import com.braintreepayments.api.cardform.R;
+import com.braintreepayments.api.test.TestActivity;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.Robolectric;
+import org.robolectric.RobolectricTestRunner;
+import org.robolectric.RuntimeEnvironment;
+
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertFalse;
+import static junit.framework.Assert.assertTrue;
+
+@RunWith(RobolectricTestRunner.class)
+public class PostalCodeEditTextTest {
+
+    private PostalCodeEditText mView;
+
+    @Before
+    public void setup() {
+        mView = Robolectric.setupActivity(TestActivity.class)
+                .findViewById(R.id.bt_card_form_postal_code);
+    }
+
+    @Test
+    public void invalidIfEmpty() {
+        assertFalse(mView.isValid());
+    }
+
+    @Test
+    public void validIfNotEmpty() {
+        mView.setText("12345");
+        assertTrue(mView.isValid());
+    }
+
+    @Test
+    public void validIfEmptyAndOptional() {
+        mView.setOptional(true);
+
+        assertTrue(mView.isValid());
+    }
+
+    @Test
+    public void getErrorMessage_returnsErrorMessageWhenEmpty() {
+        assertEquals(RuntimeEnvironment.application.getString(R.string.bt_postal_code_required), mView.getErrorMessage());
+    }
+}